requirements.txt
# Django and REST Framework
Django==4.2.7
djangorestframework==3.14.0

# JWT Authentication
djangorestframework-simplejwt==5.3.0

# PostgreSQL Database
psycopg2-binary==2.9.7

# Environment Variables
python-decouple==3.8

# CORS Headers (for frontend integration)
django-cors-headers==4.3.1

# Development tools
django-extensions==3.2.3
